+++ src/rics/mapping/__init__.py
@@ -1,7 +1,15 @@
 """Mapping implementations for matching groups of elements."""
 
-from rics.mapping import score_functions
+from rics.mapping import filter_functions, heuristic_functions, score_functions
 from rics.mapping._directional_mapping import DirectionalMapping
+from rics.mapping._heuristic_score import HeuristicScore
 from rics.mapping._mapper import Mapper
 
-__all__ = ["DirectionalMapping", "Mapper", "score_functions"]
+__all__ = [
+    "HeuristicScore",
+    "DirectionalMapping",
+    "Mapper",
+    "score_functions",
+    "filter_functions",
+    "heuristic_functions",
+]