+++ tests/mapping/test_score_functions.py
@@ -1,6 +1,5 @@
 import pandas as pd
 import pytest
-from numpy import inf
 from numpy.random import choice, randint
 
 from rics.mapping import score_functions as sf
@@ -12,55 +11,6 @@ with open("tests/mapping/words.txt") as f:
 def test_from_bad_name():
     with pytest.raises(ValueError):
         sf.from_name("does-not-exist")
-
-
-@pytest.mark.parametrize(
-    "value, add_source, expected",
-    [
-        ("id", False, [False, False, False, False, False]),
-        ("id", True, [True, False, False, False, False]),
-        ("name", False, [False, False, False, True, True]),
-        ("exact", True, [False, True, False, False, False]),
-    ],
-)
-def test_score_with_heuristics(value, add_source, expected):
-    columns = ["animal_id", "exact", "exact_animal", "cute_name", "name_real"]
-
-    fstrings = ["{value}_real", "cute_{value}"]
-    if add_source:
-        fstrings.append("{source}_{value}")
-
-    actual = list(
-        sf.score_with_heuristics(
-            value,
-            columns,
-            source="animal",  # Given by mapper
-            # Configured
-            fstrings=fstrings,
-        )
-    )
-    actual_above_05 = [v > 0.5 for v in actual]
-    assert actual_above_05 == expected
-
-
-@pytest.mark.parametrize(
-    "value, expected",
-    [
-        ("cand", [1.0, 0.0, 0.0, 0.0]),
-        ("cand0", [-inf, inf, -inf, inf]),
-        ("cand2", [-inf, inf, inf, inf]),
-    ],
-    # ids=["no-cs-1", "no-cs-2", "kw1-cs", "kw2-cs", "kw1/kw2-cs"],
-)
-def test_score_with_heuristics_short_circuiting(value, expected):
-    kwargs = {
-        "cand0": [".*kw1.*"],
-        "cand2": [".*kw1.*", ".*kw2.*"],
-    }
-    candidates = ["cand", "cand-kW1", "cand-KW2", "cand-kW1-kw2"]
-
-    actual = list(sf.score_with_heuristics(value, candidates, **kwargs))
-    assert actual == expected
 
 
 @pytest.mark.parametrize(
@@ -76,7 +26,7 @@ def test_like_database_table(value, expected_max_table):
     candidate_tables = ["humans", "animals"]
     score = sf.from_name("like_database_table")
 
-    s = pd.Series(score(value, candidate_tables), index=candidate_tables)
+    s = pd.Series(score(value, candidate_tables, None), index=candidate_tables)
     assert s.idxmax() == expected_max_table, s
 
 
@@ -88,13 +38,11 @@ def test_stable(func, dtype):
     candidates = make(12, dtype)
     values = make(6, dtype)
 
-    kwargs = {"fstrings": ["un_{value}"]} if func == sf.score_with_heuristics else {}
-
     for v in values:
-        actual_scores = list(func(v, candidates.copy(), **kwargs))
+        actual_scores = list(func(v, candidates.copy(), None))
         assert all([isinstance(s, float) for s in actual_scores]), "Bad return type"
 
-        expected_scores = [next(iter(func(v, [candidates[i]], **kwargs))) for i in range(len(candidates))]
+        expected_scores = [next(iter(func(v, [candidates[i]], None))) for i in range(len(candidates))]
         assert actual_scores == expected_scores